+++ src/form/form-group.component.ts
@@ -8,6 +8,8 @@ const internalIconMap = {
     date: 'wtf wtf-schedule-o'
 };
 
+type TipsMode = 'default' | 'label';
+
 @Component({
     selector: 'thy-form-group',
     templateUrl: './form-group.component.html',
@@ -21,6 +23,7 @@ export class ThyFormGroupComponent implements OnInit {
     feedbackIcon: string;
     feedbackSvgIconName: string;
     tips: string;
+    tipMode: TipsMode = 'default';
 
     @HostBinding('class.row-fill') _rowFill = false;
 
@@ -64,6 +67,11 @@ export class ThyFormGroupComponent implements OnInit {
             this.feedbackSvgIconName = value;
             this.feedbackIcon = null;
         }
+    }
+
+    @Input()
+    set thyTipsMode(mode: TipsMode) {
+        this.tipMode = mode;
     }
 
     @Input()