+++ src/rate/examples/count/count.component.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-rate-count-example',
+    templateUrl: './count.component.html'
+})
+export class ThyRateCountExampleComponent implements OnInit {
+    value = 2;
+
+    constructor() {}
+
+    ngOnInit() {}
+}