+++ src/rate/examples/disabled/disabled.component.ts
@@ -1,0 +1,13 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-rate-disabled-example',
+    templateUrl: './disabled.component.html'
+})
+export class ThyRateDisabledExampleComponent implements OnInit {
+    value = 2;
+
+    constructor() {}
+
+    ngOnInit() {}
+}